refactor(main): split monolith into api/schemas, persistence, read_routes, admin_routes
Extract main.py (~1450 lines) into focused modules:

- src/persistence.py: file-path constants and data-loading helpers
- src/api/schemas.py: request/response Pydantic models
- src/api/read_routes.py: shared read-only APIRouter (9 GET endpoints)
- src/api/admin_routes.py: mutating admin endpoints (POST/DELETE/PATCH/GET-export)

main.py retains only app construction, CORS, static/template mounting,
router inclusion, the two HTML routes, and the __main__ runner (~170 lines).

All test mock paths updated to target the new module namespaces.
Admin routes reference persistence constants via module alias
(from src import persistence as _persistence) so integration tests
can patch a single location for both reads and writes.

No API surface changes -- OpenAPI spec is identical.
